feat: split up pixi manifest lib
