+++ crates/pixi_manifest/src/target.rs
@@ -561,7 +561,7 @@ mod tests {
     use insta::assert_snapshot;
     use itertools::Itertools;
 
-    use crate::Manifest;
+    use crate::manifest::Manifest;
 
     #[test]
     fn test_targets_overwrite_order() {
